Flatten some common HttpUrl internals
Trying to reduce the number of indirections that were only
necessary when we were attempting to support multiplatform.
