+++ okhttp/src/main/kotlin/okhttp3/internal/http/StatusLine.kt
@@ -16,9 +16,8 @@
  */
 package okhttp3.internal.http
 
-import kotlin.jvm.JvmField
+import java.net.ProtocolException
 import okhttp3.Protocol
-import okhttp3.ProtocolException
 import okhttp3.Response
 import okio.IOException
